+++ src/main/java/org/purpura/apimg/controller/empresa/oas/ResiduoContract.java
@@ -125,7 +125,7 @@ public interface ResiduoContract {
     ResiduoResponseDTO downturnResiduo(@Parameter(description = "CNPJ da empresa", example = "12345678000195") @PathVariable String cnpj,
                        @Parameter(description = "ID do resíduo", example = "1") @PathVariable String id,
                        @Parameter(description = "Quantidade a ser baixada", example = "10")
-                       @Min(value = 1, message = "A quantidade da baixa deve ser no mínimo 1 unidade.") @RequestParam Integer quantity);
+                       @Min(value = 1, message = "A quantidade da baixa deve ser no mínimo 1 unidade.") @RequestParam Long quantity);
 
 
 }
