+++ src/main/java/org/purpura/apimg/exception/empresa/ResiduoInsufficientStockException.java
@@ -1,7 +1,7 @@
 package org.purpura.apimg.exception.empresa;
 
 public class ResiduoInsufficientStockException extends RuntimeException {
-    public ResiduoInsufficientStockException(String idResiduo, Integer quantidade) {
+    public ResiduoInsufficientStockException(String idResiduo, Long quantidade) {
         super(String.format("Não há estoque suficiente para o resíduo com id %s. Quantidade solicitada: %d", idResiduo, quantidade));
     }
 }
